+++ src/lib/connection/XMPPService.ts
@@ -1043,9 +1043,9 @@ class XMPPService extends GenericService {
                 "xml:lang": lang
             }, message), xml("request", {
                     "xmlns": NameSpacesLabels.ReceiptsNameSpace
-                }, xml("active", {
+            }), xml("active", {
                     "xmlns": NameSpacesLabels.ChatestatesNameSpace
-                })
+                }
             ));
 
             if (that.copyMessage == false) {
@@ -1087,7 +1087,7 @@ class XMPPService extends GenericService {
             return new Promise((resolve, reject) => {
                 that.xmppClient.send(stanza).then(() => {
                     that.logger.log("debug", LOG_ID + "(sendChatMessage) sent");
-                    resolve({from: that.jid_im, to: jid, type: "chat", id: id, date: new Date(), content: message});
+                    resolve({from: that.jid_im, to: jid, lang: lang, type: "chat", id: id, date: new Date(), content: message, urgency: urgency});
                 }).catch((err) => {
                     return reject(err);
                 });
@@ -1235,19 +1235,29 @@ class XMPPService extends GenericService {
                 if (data === "") {
                     xmppMessage = xml("message", {to: to, type: "chat", id: origMsgId, "xml:lang": lang},
                             xml("body", {"xml:lang": lang}, data),
-                            xml("delete", {"xmlns": NameSpacesLabels.MessageCorrectNameSpace}),
+                            xml("deleted", {"xmlns": NameSpacesLabels.MessageCorrectNameSpace}),
                             xml("store", {"xmlns": NameSpacesLabels.HintsNameSpace}),
                             xml("request", {"xmlns": NameSpacesLabels.ReceiptNS}),
                             xml("active", {"xmlns": NameSpacesLabels.ChatstatesNS})
                     );
                 } else {
-                    xmppMessage = xml("message", {to: to, type: "chat", id: origMsgId, "xml:lang": lang},
-                            xml("body", {"xml:lang": lang}, data),
-                            xml("modify", {"xmlns": NameSpacesLabels.MessageCorrectNameSpace}),
-                            xml("store", {"xmlns": NameSpacesLabels.HintsNameSpace}),
-                            xml("request", {"xmlns": NameSpacesLabels.ReceiptNS}),
-                            xml("active", {"xmlns": NameSpacesLabels.ChatstatesNS})
-                    );
+                    if (data == undefined) {
+                        xmppMessage = xml("message", {to: to, type: "chat", id: origMsgId, "xml:lang": lang},
+                                //xml("body", {"xml:lang": lang}, data),
+                                xml("modified", {"xmlns": NameSpacesLabels.MessageCorrectNameSpace}),
+                                xml("store", {"xmlns": NameSpacesLabels.HintsNameSpace}),
+                                xml("request", {"xmlns": NameSpacesLabels.ReceiptNS}),
+                                xml("active", {"xmlns": NameSpacesLabels.ChatstatesNS})
+                        );
+                    } else {
+                        xmppMessage = xml("message", {to: to, type: "chat", id: origMsgId, "xml:lang": lang},
+                                xml("body", {"xml:lang": lang}, data),
+                                xml("modified", {"xmlns": NameSpacesLabels.MessageCorrectNameSpace}),
+                                xml("store", {"xmlns": NameSpacesLabels.HintsNameSpace}),
+                                xml("request", {"xmlns": NameSpacesLabels.ReceiptNS}),
+                                xml("active", {"xmlns": NameSpacesLabels.ChatstatesNS})
+                        );
+                    }
                 }
             }
 
@@ -1279,17 +1289,34 @@ class XMPPService extends GenericService {
                 if (data==="") {
                     xmppMessage = xml("message", {to: conversation.bubble.jid, type: "groupchat", id: origMsgId},
                             xml("body", {"xml:lang": lang}, data),
-                            xml("delete", {"xmlns": NameSpacesLabels.MessageCorrectNameSpace}),
+                            xml("deleted", {"xmlns": NameSpacesLabels.MessageCorrectNameSpace}),
                             xml("store", {"xmlns": NameSpacesLabels.HintsNameSpace}),
                             xml("request", {"xmlns": NameSpacesLabels.ReceiptNS}),
                             xml("active", {"xmlns": NameSpacesLabels.ChatstatesNS}));
                 } else {
-                    xmppMessage = xml("message", {to: conversation.bubble.jid, type: "groupchat", id: messageToSendID},
-                            xml("body", {"xml:lang": lang}, data),
-                            xml("modify", {"xmlns": NameSpacesLabels.MessageCorrectNameSpace}),
-                            xml("store", {"xmlns": NameSpacesLabels.HintsNameSpace}),
-                            xml("request", {"xmlns": NameSpacesLabels.ReceiptNS}),
-                            xml("active", {"xmlns": NameSpacesLabels.ChatstatesNS}));                    
+                    if (data == undefined) {
+                        xmppMessage = xml("message", {
+                                    to: conversation.bubble.jid,
+                                    type: "groupchat",
+                                    id: messageToSendID
+                                },
+                                //xml("body", {"xml:lang": lang}, data),
+                                xml("modified", {"xmlns": NameSpacesLabels.MessageCorrectNameSpace}),
+                                xml("store", {"xmlns": NameSpacesLabels.HintsNameSpace}),
+                                xml("request", {"xmlns": NameSpacesLabels.ReceiptNS}),
+                                xml("active", {"xmlns": NameSpacesLabels.ChatstatesNS}));
+                    } else {
+                        xmppMessage = xml("message", {
+                                    to: conversation.bubble.jid,
+                                    type: "groupchat",
+                                    id: messageToSendID
+                                },
+                                xml("body", {"xml:lang": lang}, data),
+                                xml("modified", {"xmlns": NameSpacesLabels.MessageCorrectNameSpace}),
+                                xml("store", {"xmlns": NameSpacesLabels.HintsNameSpace}),
+                                xml("request", {"xmlns": NameSpacesLabels.ReceiptNS}),
+                                xml("active", {"xmlns": NameSpacesLabels.ChatstatesNS}));
+                    }
                 }
             }
             if (that.copyMessage==false) {